# openapi.yaml
openapi: 3.1.0
info:
  title: SsuBench API
  version: "1.0"
servers:
- url: http://localhost:8080
security:
- bearerAuth: []
paths:
  /api/tasks:
    get:
      tags:
      - task-controller
      summary: Получить все задачи
      operationId: getAllTasks
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelTaskResponse"
    post:
      tags:
      - task-controller
      summary: Создать задачу
      operationId: createTask
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskResponse"
  /api/tasks/{taskId}/select-executor/{bidId}:
    post:
      tags:
      - task-controller
      summary: Выбрать исполнителя
      operationId: selectExecutor
      parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: bidId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskResponse"
  /api/tasks/{taskId}/reject:
    post:
      tags:
      - task-controller
      summary: Отправить задачу на доработку
      operationId: rejectTask
      parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskResponse"
  /api/tasks/{taskId}/confirm:
    post:
      tags:
      - task-controller
      summary: Подтвердить выполнение задачи
      operationId: confirmTask
      parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskResponse"
  /api/tasks/{taskId}/complete:
    post:
      tags:
      - task-controller
      summary: Выполнить задачу
      operationId: completeTask
      parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskResponse"
  /api/tasks/{taskId}/cancel:
    post:
      tags:
      - task-controller
      summary: Отменить задачу
      operationId: cancelTask
      parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskResponse"
  /api/tasks/{taskId}/bids:
    get:
      tags:
      - task-controller
      summary: Получить все отклики на задачу
      operationId: getTaskBids
      parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BidResponse"
    post:
      tags:
      - task-controller
      summary: Откликнуться на задачу
      operationId: createBid
      parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BidResponse"
  /api/auth/register:
    post:
      tags:
      - auth-controller
      summary: Зарегистрироваться
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/auth/login:
    post:
      tags:
      - auth-controller
      summary: Войти
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/admin/users/block:
    post:
      tags:
      - admin-controller
      summary: Заблокировать/разблокировать пользователя
      operationId: blockUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockUserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/users/me:
    get:
      tags:
      - user-controller
      summary: Получить информацию о текущем пользователе
      operationId: getProfile
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
  /api/tasks/{taskId}:
    get:
      tags:
      - task-controller
      summary: Получить задачу по id
      operationId: getTaskById
      parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TaskResponse"
  /api/tasks/my-assignments:
    get:
      tags:
      - task-controller
      summary: Получить мои задачи (для исполнителя)
      operationId: getMyAssignedTasks
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelTaskResponse"
components:
  schemas:
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        cost:
          type: integer
          format: int64
          minimum: 1
      required:
      - cost
      - description
      - title
    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        cost:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - CREATED
          - IN_PROGRESS
          - COMPLETED
          - CONFIRMED
          - CANCELLED
        customerUsername:
          type: string
        executorUsername:
          type: string
        description:
          type: string
    BidResponse:
      type: object
      properties:
        bidId:
          type: integer
          format: int64
        taskId:
          type: integer
          format: int64
        executorUsername:
          type: string
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          maxLength: 100
          minLength: 6
        role:
          type: string
          enum:
          - CUSTOMER
          - EXECUTOR
          - ADMIN
      required:
      - role
      - username
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
      required:
      - password
      - username
    BlockUserRequest:
      type: object
      properties:
        username:
          type: string
        blocked:
          type: boolean
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        points:
          type: integer
          format: int64
        role:
          type: string
          enum:
          - CUSTOMER
          - EXECUTOR
          - ADMIN
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
    PagedModelTaskResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/TaskResponse"
        page:
          $ref: "#/components/schemas/PageMetadata"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT